removed / from the directory starting inside manifest.json
I was getting an error where my deployed webpage would simply show blank page. Hence, I removed forward-slash(/) from the index.html src and on the manifest.json. It actually worked, but not I am getting another error where my images are not loading as they couldn't be found on the source directory so I am trying by changing the directory path ( that is simply removing (/) in the starting of every file path.